Significantly improve performance of detected dynamic libs
Before the routine would extract each line, one by one, from the command
result block, then search each line for "NEEDED" and print if found.
Now it searches it on a block level and leaves when no more or found.
E.g. 10 searches version 200 or 2000 (depending on symbols in
executable)