Add support for GCC runtime
There's a chicken/egg problem with GCC runtime library support.
The solution is to keep the compiler subpackages with the system root.
When ravenadm is executed, it looks for missing compiler subpackages in
the packages directory and copies over all missing ones from the
sysroot.
Then, if a compiler subpackage is specified as an "extra-run" depends,
get the package version from the definitions.  The port tree can't be
used because
A) The port very well might not have been scanned so the info is missing
and
B) the port version of gcc might differ from the sysroot version.
The A) case actually causes pkg(8) to segfault!