fix full tree scan and (so far unused) custom directory scan
The custom ports are not additional to, but instead of.  Most of them
replace official ports.  Therefore the custom scan must set additional
port key entries to custom, not add a new entry.

On the full scan, we need to populate the portslist as if the user
submitted a list of every port via the command line.  This unbreaks at
least delete-package functionality (which is based on portslist)