Major change to unkindness functionality
It turns out that construction source ports as needed will not work.
The compiled buildsheet is required for all ports that used DEF
(definitions) which is basically all ports.

Thus, add logic to determine if custom buildsheets are available and
current, and if a custom unkindness_variants file is available and
current, and generate everything that's not.

The next step is to revert the recent commits (mostly) to use these new
buildsheets and indices.