Add ABI support for MacOS
Right now half the ABI is hardcoded because pkg(8) is returning the
wrong value.  So I've hardcoded the ABI to match.  MacOS doesn't use ELF
format, so there is no ELF notes, and thus the OS version is not coded
into /bin/sh.

Passing the release to the funcion makes me ask why we're reading the
notes in any case.  Since we already determined all this for the slave
environment, why not just keep using that?