Add new "jump" dev subcommand
The command is for use with shell's "eval" command so that the developer
can switch the ports subdirectory with a single command, e.g.
eval `ravenadm dev jump libreoffice`

Which here, fix a bug in the "locate" command -- it should a) have
searched customsource first and B) it needs to seach the generated
buildsheets directory, not the source directory, in order to match
standard ports.