Fix unhandled  exception on FreeBSD build box
Since I reused the port container which has a current maximum of 8000 items,
the purge of obsolete package operation fails because there are more than
8000 packages in the repo.
The solution is to create a new container type, one with a higher number of
allowed items.  Currently this new container can hold 24,000 packages.

Bump revision.