add support for libmap inside builder
DragonFly, FreeBSD and MidnightBSD have rtld that support libmap.
This change allows a profile-specific libmap.conf to be installed in the
builder.  The initial need for this is to bootstrap MidnightBSD which has
a slightly different libc.  libmap will allow the builder's programs to use
FreeBSD's libc while the generated programs will use the native
MidnightBSD libc.