Attempt to support subpackages with EXRUN (#70)
We need to support e.g.
EXRUN[tools]= primary
When there are multiple variants, the existing method can't pull in the proper subpackage.

While here, improve the EXRUN transformation.  Before it bruteforce checked the entire array for each possible value.
Now it should iterate the arrow once and check each value to see if it's special.