Change order of deinstall
Don't deinstall packages until the target package is built.

Otherwise rvn can't find the required libraris on the runpaths because
they've already been removed.

Plus, we don't want to remove packages with the "build" command, only the "test" command.

Ravenadm has been like this for a long time and it was probably a mistake