Kill 5 year old curses bug
I've been chasing this down for YEARS.
I suspected a constraint issue was the problem but since it never got logged I stopped looking for it.

One the logging was fixed (a flush was needed .. the error never made it to the log file) the problem and fix quickly followed.