Fix long standing bug (requires testing)
These four modules were validated to take exactly one known subpackage:
schema, mime-info, gnome-icons, desktop-utils.
However, the code was attaching the dependencies to the primary package, not the one specified in the argument.

This code extracts the argument (trusted to be a known single subpackage
name) and sets exrun depends of subpackage rather than primary run-depends.