Add hack to deal with freezing on Midnight and FreeBSD
The display freezing continues on FreeBSD and Midnight despite
my best efforts.  I can't explain why this is happening.
The theory is the SIO.Read is never returning on ZFS systems.

This didn't happen when wc was used to count line, so for now, go
back to that approach on FreeBSD and midnight.

If the freezing stops, at least it's confirmed the issue is with FOP.lines_in_log()